Add `ScatterWorkflowStep` and `LoopWorkflowStep`
To prepare the library for the new `ScatterWorkflowStep` and
`LoopWorkflowStep` objects introduced in CWL v1.3, this commit
adds type aliases to be used in previous CWL versions.

The `ScatterWorkflowStep` is an alias for `WorkflowStep` in CWL up to
v1.2, as the `WorkflowStep` object has the `scatter` and `scatterMethod`
properties.

The `LoopWorkflowStep` does not exist before CWL 1.3. Therefore, a
`NoType` abstract class has been introduced as a placeholder.